Matrix subdivision bugfix
Whenever the matrix size was not evenly divisible by the thread count,
some rows at the very end of the matrix could be left out of the
computation. This is now fixed by evenly dividing the remainder of rows
over the first threads and adjusting lower and upper bounds accordingly.